# docs/swagger.yaml
openapi: 3.0.0
info:
  title: ULenguage Backend API
  version: 1.0.0
  description: API para traducción quechua - turistas en Cusco
  contact:
    name: Equipo ULenguage
    email: soporte@ulenguage.com
servers:
  - url: http://localhost:5000
    description: Servidor local de desarrollo

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: "60d21b4667d0d8992e610c85"
        name:
          type: string
          example: "Ana García"
        email:
          type: string
          format: email
          example: "ana@mail.com"
        plan:
          type: string
          enum: [free, premium]
          example: "free"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    UserAuth:
      type: object
      properties:
        _id:
          type: string
          example: "60d21b4667d0d8992e610c85"
        name:
          type: string
          example: "Ana García"
        email:
          type: string
          format: email
          example: "ana@mail.com"
        plan:
          type: string
          enum: [free, premium]
          example: "free"
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    
    Plan:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
          example: "Premium"
        description:
          type: string
          example: "Plan completo para exploradores culturales"
        price:
          type: number
          example: 5.99
        features:
          type: array
          items:
            type: string
          example: ["OCR ilimitado", "Audio pronunciación", "Sin anuncios"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    Content:
      type: object
      properties:
        _id:
          type: string
        term:
          type: string
          example: "Allin p'unchay"
        translationEs:
          type: string
          example: "Buenos días"
        translationEn:
          type: string
          example: "Good morning"
        context:
          type: string
          example: "Se usa hasta el mediodía en comunidades rurales"
        pronunciation:
          type: string
          example: "AH-lyeen POON-chay"
        category:
          type: string
          example: "saludos"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Error interno del servidor"

tags:
  - name: auth
    description: Endpoints de autenticación
  - name: plans
    description: Gestión de planes de suscripción
  - name: content
    description: Contenido cultural quechua
  - name: seed
    description: Población de datos iniciales
  - name: ocr
    description: OCR y análisis de imágenes con IA

paths:
  /api/auth/register:
    post:
      tags: [auth]
      summary: Registro de usuario
      description: Crear una nueva cuenta de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: "Ana García"
                email:
                  type: string
                  format: email
                  example: "ana@mail.com"
                password:
                  type: string
                  minLength: 6
                  example: "123456"
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuth'
        '400':
          description: Usuario ya existe o datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/login:
    post:
      tags: [auth]
      summary: Inicio de sesión
      description: Autenticar usuario con email y contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "ana@mail.com"
                password:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuth'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/planes:
    get:
      tags: [plans]
      summary: Listar planes disponibles
      description: Obtener todos los planes de suscripción
      responses:
        '200':
          description: Lista de planes obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/seed:
    post:
      tags: [seed]
      summary: Ejecutar seeders
      description: Poblar la base de datos con datos iniciales (planes y contenido cultural)
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Seeders ejecutados exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Seeders ejecutados correctamente"
                  success:
                    type: boolean
                    example: true
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/news:
    get:
      tags: [news]
      summary: Obtener últimas noticias culturales
      description: Devuelve las últimas noticias filtradas por idioma (español, inglés o quechua)
      parameters:
        - in: query
          name: lang
          schema:
            type: string
            enum: [es, en, qu]
            default: es
          description: Idioma de las noticias (es=español, en=inglés, qu=quechua)
        - in: query
          name: limit
          schema:
            type: integer
            default: 3
            minimum: 1
            maximum: 10
          description: Número máximo de noticias a devolver
      responses:
        '200':
          description: Noticias obtenidas exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 3
                  language:
                    type: string
                    example: "es"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          example: "60d21b4667d0d8992e610c85"
                        title:
                          type: string
                          example: "Inti Raymi 2025: La Fiesta del Sol vuelve a Sacsayhuamán"
                        content:
                          type: string
                          example: "La tradicional Fiesta del Sol..."
                        summary:
                          type: string
                          example: "El Inti Raymi 2025 celebrará..."
                        imageUrl:
                          type: string
                          example: "https://images.unsplash.com/photo-1587595431973-160d0d94add1?w=800"
                        language:
                          type: string
                          enum: [es, en, qu]
                          example: "es"
                        category:
                          type: string
                          example: "festividad"
                        isActive:
                          type: boolean
                          example: true
                        publishedAt:
                          type: string
                          format: date-time
                          example: "2025-06-15T00:00:00.000Z"
                        author:
                          type: string
                          example: "Equipo ULenguage"
                        viewCount:
                          type: number
                          example: 0
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/ocr/status:
    get:
      tags: [ocr]
      summary: Estado del servicio OCR
      description: Verifica la disponibilidad del servicio OCR y Google Vision API
      responses:
        '200':
          description: Servicio OCR disponible y funcional
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  service:
                    type: string
                    example: "OCR Vision AI"
                  provider:
                    type: string
                    example: "Google Cloud Vision"
                  ready:
                    type: boolean
                    example: true
                  uploadsDir:
                    type: boolean
                    example: true
                  timestamp:
                    type: string
                    format: date-time
        '503':
          description: Servicio OCR no disponible
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: false
                  service:
                    type: string
                    example: "OCR Vision AI"
                  ready:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Servicio no disponible"
                  code:
                    type: string
                    example: "SERVICE_UNAVAILABLE"

  /api/ocr/analyze:
    post:
      tags: [ocr]
      summary: Analizar imagen con OCR y contexto cultural
      description: Procesa una imagen con Google Vision, detecta texto y objetos, y genera explicaciones culturales usando IA
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: Imagen a analizar (JPG, PNG o WEBP, máx 5MB)
                targetLang:
                  type: string
                  enum: [es, en, qu]
                  default: es
                  description: Idioma objetivo para las explicaciones
                langs:
                  type: string
                  example: "es,en,qu"
                  description: Lista de idiomas separados por comas para generar explicaciones
      responses:
        '200':
          description: Imagen analizada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  texts:
                    type: array
                    items:
                      type: object
                      properties:
                        text:
                          type: string
                        explanations:
                          type: object
                          additionalProperties:
                            type: string
                        explanationsPreview:
                          type: object
                          additionalProperties:
                            type: string
                  detectedLang:
                    type: string
                    example: "es"
                  labels:
                    type: array
                    items:
                      type: string
                  objects:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        score:
                          type: number
                          format: float
                        boundingBox:
                          type: array
                          items:
                            type: object
                        translatedNames:
                          type: object
                          additionalProperties:
                            type: string
                        explanations:
                          type: object
                          additionalProperties:
                            type: string
                        explanationsPreview:
                          type: object
                          additionalProperties:
                            type: string
                  explanationProvidedLangs:
                    type: array
                    items:
                      type: string
                  requestedLang:
                    type: string
        '400':
          description: Error en la solicitud (archivo faltante o inválido)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  code:
                    type: string
                    enum: [NO_FILE, INVALID_FILE_TYPE, TOO_MANY_FILES]
        '413':
          description: Archivo demasiado grande
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "La imagen es demasiado grande. Tamaño máximo: 5 MB."
                  code:
                    type: string
                    example: "FILE_TOO_LARGE"
                  maxSize:
                    type: string
                    example: "5 MB"
        '500':
          description: Error al procesar la imagen
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  code:
                    type: string
                    enum: [OCR_ERROR, VISION_API_ERROR, TIMEOUT, AUTH_ERROR]
        '504':
          description: Timeout al procesar imagen
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  code:
                    type: string
                    example: "TIMEOUT"

  /api/ocr/analyze-and-translate:
    post:
      tags: [ocr]
      summary: Analizar y traducir imagen
      description: Procesa una imagen con OCR, detecta objetos y genera explicaciones culturales traducidas
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: Imagen a analizar (JPG, PNG o WEBP, máx 5MB)
                targetLang:
                  type: string
                  enum: [es, en, qu]
                  default: es
                  description: Idioma objetivo
                langs:
                  type: string
                  example: "es,en,qu"
      responses:
        '200':
          description: Imagen analizada y traducida exitosamente
        '400':
          description: Error en la solicitud
        '413':
          description: Archivo demasiado grande
        '500':
          description: Error al procesar